Updated runtime code to compile with gcc 5.3.0, see #4943
git-svn-id: https://svn.jmodelica.org@8832 7ae51c65-13d5-4921-9d80-6511ddd1419c

+++ trunk/RuntimeLibrary/src/jmi/jmi_log.c
@@ -638,7 +638,7 @@ static int emitted_category(log_t *log, category_t c) {
     }
 
     /* interim solution to allow to pass a log level >= 4 instead of a category */
-    if (c >= 4) {
+    if (((int)c) >= 4) {
         if (cb->log_options.log_level < c) return FALSE;
         c = logInfo;
     }
